requirements.txt
requests>=2.28
python-dotenv
azure-ai-documentintelligence
